src/requirements/prd.txt
Jinja2==2.11.2
python-dotenv==0.13.0
requests==2.24.0 # from 2.23.0
tqdm==4.46.1
